# swagger.yaml
openapi: 3.0.0
info:
  title: HDR UK API 
  description: API for Tools and artefacts repository.
  version: 0.0.1
servers:
  - url: https://api.www.healthdatagateway.org/api
  - url: http://localhost:3001/api

  - url: https://api.{environment}.healthdatagateway.org:{port}/api
    variables:
      environment:
        default: latest
        description: The Environment name.
      port:
        enum:
          - '443'
        default: '443'
paths:
...
  /v1/datasets/:
    get:
      summary: Returns List of Dataset objects.
      tags:
        - Datasets
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/person/:
    get:
      summary: Returns an array of person objects.
      tags:
        - Person
      responses:
        '200':
          description: OK
    post:
      summary: Returns a new person object.
      tags:
        - Person
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - firstname
                - lastname
                - bio
                - link
                - orcid
                - emailNotifications
                - terms
              properties:
                firstname:
                  type: string
                lastname:
                  type: string
                bio:
                  type: string
                link:
                  type: string
                orcid:
                  type: string
                emailNotifications:
                  type: boolean
                terms:
                  type: boolean
              example:
                firstname: 'John'
                lastname: 'Smith'
                bio: 'Researcher'
                link: 'http://google.com'
                orcid: 'https://orcid.org/123456789'
                emailNotifications: false
                terms: true
      responses:
        '200':
          description: OK
    put:
      summary: Returns edited person object.
      tags:
        - Person
      responses:
        '200':
          description: OK
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - id
                - bio
                - link
                - orcid
                - emailNotifications
                - terms
              properties:
                id:
                  type: string
                bio:
                  type: string
                link:
                  type: string
                orcid:
                  type: string
                emailNotifications:
                  type: boolean
                terms:
                  type: boolean
              example:
                id: '5268590523943617'
                bio: 'Research assistant'
                link: 'http://google.com'
                orcid: 'https://orcid.org/123456789'
                emailNotifications: false
                terms: true
  /v1/search/:
    get:
      tags:
        - Search
      summary: Search for HDRUK /search?search
      produces:
        - application/json
      parameters:
        - in: query
          name: params
          schema:
            type: object
            properties:
              search:
                type: string
                example: Epilepsy
              type:
                type: string
                example: all
              category:
                type: string
                example: API
              programmingLanguage:
                type: string
                example: Javascript
              features:
                type: string
                example: Arbitrage
              topics:
                type: string
                example: Epilepsy
              startIndex:
                type: string
                example: 0
              maxResults:
                type: string
                example: 10
          style: form
          explode: true
      responses: 
        '200':
          description: OK
  /v1/datasets/?search={searchTerm}:
    get:
      summary: Returns dataset search results by searchTerm
      tags:
        - Datasets
      parameters:
        - name: searchTerm
          in: path
          required: true
          description: A search term.
          schema:
            type : string
            example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/stats?rank={rank}:
    get:
      summary: Returns the details on recent searches, popular objects or recently updated objects based on the rank query parameter.
      tags:
        - Stats
      parameters:
        - name: rank
          in: path
          required: true
          description: The type of stat.
          schema:
            type : string
            example: recent
      responses: 
        '200':
          description: OK
...
  /v1/stats?rank={rank}&type={type}&month={month}&year={year}:
    get:
      summary: Returns unmet demand information for the resource type specified, based on the month and year query params.
      tags:
        - Stats
      parameters:
        - name: rank
          in: path
          required: true
          description: The type of stat.
          schema:
            type : string
            example: unmet
        - name: type
          in: path
          required: true
          description: Resource type.
          schema:
            type : string
            example: Tools
        - name: month
          in: path
          required: true
          description: Month number.
          schema:
            type : string
            example: 7
        - name: year
          in: path
          required: true
          description: Year.
          schema:
            type : string
            example: 2020
      responses: 
        '200':
          description: OK
  /v1/kpis?kpi={type}:
    get:
      summary: Returns information for KPIs, based on the KPI type parameter.
      tags:
        - KPIs
      parameters:
        - name: type
          in: path
          required: true
          description: The type of KPI.
          schema:
            type : string
            example: technicalmetadata
      responses: 
        '200':
          description: OK
  /v1/kpis?kpi={type}&selectedDate={selectedDate}:
    get:
      summary: Returns information for KPIs, based on the KPI type and selectedDate parameters.
      tags:
        - KPIs
      parameters:
        - name: type
          in: path
          required: true
          description: The type of KPI.
          schema:
            type : string
            example: uptime
        - name: selectedDate
          in: path
          required: true
          description: Full date time string.
          schema:
            type : string
            example: Wed Jul 01 2020 01:00:00 GMT 0100 (British Summer Time)
      responses: 
        '200':
          description: OK
...
  /v1/messages/:
  
    post:
      summary: Returns a new Message object and creates an associated parent Topic if a Topic is not specified in request body
      security:
        - cookieAuth: []
      tags:
        - Messages
      parameters:
        - in: body 
          name: payload
          description: Payload for updating existing Message
          schema: 
            type: object
            properties:
              isRead:
                type: boolean
              messageDescription:
                type: string
              messageType:
                type: string
      responses: 
        '201':
          description: OK
  /v1/topics/:
    post:
      summary: Returns a new Topic object with ID (Does not create any associated messages)
      security:
        - cookieAuth: []
      tags:
        - Topics
      parameters:
        - in: body 
          name: payload
          description: Payload for posting new message Topic
          schema: 
            type: object
            properties:
              relatedObjectIds:
                type: array
                example: "['1','2','3']"
      responses:
        '201':
          description: A new Topic
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: object
                    description: Generated ID
                  title:
                    type: string
                    description: Title of message
                  subtitle:
                    type: string
                    description: Subtitle of message
                  relatedObjectIds:
                    type: array
                    description: Object ID this Topic is related to
                  createdBy:
                    type: object
                    description: User that created the topic
                  createdDate:
                    type: date
                    description: Date the topic was created
                  recipients:
                    type: array
                    description: Collection of user IDs
                  tags:
                    type: array
                    description: Collection of tags to describe topic
    get:
      summary: Returns a list of all topics that the authenticated user is a recipient or member of
      security:
        - cookieAuth: []
      tags:
        - Topics
      responses:
        '200':
          description: Ok                
...
  /v1/projects/:
    post:
      summary: Returns a Project object with ID.
      security:
        - cookieAuth: []
      tags:
        - Projects
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                type:
                  type: string
                name:
                  type: string
                link:
                  type: string
                description:
                  type: string
                categories:
                  type: object
                  properties:
                    category:
                      type: string
                    programmingLanguage:
                      type: array
                      items:
                        type: string
                    programmingLanguageVersion:
                      type: string
                licence:
                  type: string
                authors:
                  type: array
                  items:
                    type: number
                tags:
                  type: object
                  properties:
                    features:
                      type: array
                      items:
                        type: string
                    topics:
                      type: array
                      items:
                        type: string      
              example:   # Sample object
                type: 'project'
                name: 'Epilepsy data research'
                link: 'http://epilepsy.org'
                description: 'Epilespy data research description'
                categories: { category: 'API', programmingLanguage: ['Javascript'], programmingLanguageVersion: '0.0.0'}
                licence: 'MIT licence'
                authors: [4495285946631793]
                tags: {features: ['Arbitrage'], topics: ['Epilepsy']}
      responses: 
        '200':
          description: OK
          
    get:
      summary: Returns List of Project objects.
      tags:
        - Projects
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/papers/:
    post:
      summary: Returns a Paper object with ID.
      security:
        - cookieAuth: []
      tags:
        - Papers
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              required:
                - name
              properties:
                type:
                  type: string
                name:
                  type: string
                link:
                  type: string
                description:
                  type: string
                categories:
                  type: object
                  properties:
                    category:
                      type: string
                    programmingLanguage:
                      type: array
                      items:
                        type: string
                    programmingLanguageVersion:
                      type: string
                licence:
                  type: string
                authors:
                  type: array
                  items:
                    type: number
                tags:
                  type: object
                  properties:
                    features:
                      type: array
                      items:
                        type: string
                    topics:
                      type: array
                      items:
                        type: string      
              example:   # Sample object
                type: 'paper'
                name: 'Epilepsy data research'
                link: 'http://epilepsy.org'
                description: 'Epilespy data research description'
                categories: { category: 'API', programmingLanguage: ['Javascript'], programmingLanguageVersion: '0.0.0'}
                licence: 'MIT licence'
                authors: [4495285946631793]
                tags: {features: ['Arbitrage'], topics: ['Epilepsy']}
      responses: 
        '200':
          description: OK
    get:
      summary: Return List of Paper objects.
      tags:
        - Papers
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
...
  /v1/tools/:
    get:
      summary: Return List of Tool objects.
      tags:
        - Tools
      parameters:
      - in: query
        name: limit
        required: false
        description: Limit the number of results
        schema:
          type : integer
          example: 3
      - in: query
        name: offset
        required: false
        description: Index to offset the search results
        schema:
          type : integer
          example: 1
      - in: query
        name: q
        required: false
        description: Filter using search query
        schema:
          type : string
          example: epilepsy
      responses: 
        '200':
          description: OK
...